requirements.txt
argcomplete==3.6.2
asgiref==3.8.1
awsebcli==3.24.1
bcrypt==4.3.0
blessed==1.21.0
botocore==1.38.19
cement==2.10.14
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==45.0.2
decorator==5.2.1
Deprecated==1.2.18
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
fabric==3.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
invoke==2.2.0
jmespath==1.0.1
packaging==25.0
paramiko==3.5.1
pathspec==0.12.1
pipx==1.7.1
platformdirs==4.3.8
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyNaCl==1.5.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
uritemplate==4.1.1
urllib3==1.26.20
userpath==1.9.2
wcwidth==0.2.13
whitenoise==6.9.0
wrapt==1.17.2
